Be145

+++ app/Http/Controllers/AdminAnimeController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\AdminAnime;
+use App\Models\ChucVu;
+use App\Models\PhanQuyen;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -14,28 +16,59 @@ class AdminAnimeController extends Controller
 {
     public function getData()
     {
-        $data   = AdminAnime::select('admin_animes.*')
-                         ->get(); // get là ra 1 danh sách
+        $id_chuc_nang = 1;
+        $user   = Auth::guard('sanctum')->user(); // Chính là người đang login
+        $user_chuc_vu   = $user->id_chuc_vu;    // Giả sử
+        $check  = PhanQuyen::where('id_chuc_vu', $user_chuc_vu)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if (!$check) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
+        $chuc_vu_admin   = ChucVu::select('chuc_vus.*')
+            ->get(); // get là ra 1  sách
+        $dataAdmin   = AdminAnime::join('chuc_vus', 'id_chuc_vu', 'chuc_vus.id')
+            ->select('admin_animes.*', 'chuc_vus.ten_chuc_vu')
+            ->paginate(6); // get là ra 1  sách
+        $response = [
+            'pagination' => [
+                'total' => $dataAdmin->total(),
+                'per_page' => $dataAdmin->perPage(),
+                'current_page' => $dataAdmin->currentPage(),
+                'last_page' => $dataAdmin->lastPage(),
+                'from' => $dataAdmin->firstItem(),
+                'to' => $dataAdmin->lastItem()
+            ],
+            'dataAdmin' => $dataAdmin
+        ];
         return response()->json([
-            'admin'  =>  $data,
+            'admin'  =>  $response,
+            'chuc_vu_admin'  =>  $chuc_vu_admin,
         ]);
     }
-    public function getDataProfile(Request $request){
-        $user = AdminAnime::where('id',$request->id_admin)->first();
+    public function getDataProfile(Request $request)
+    {
+        $user = AdminAnime::join('chuc_vus', 'id_chuc_vu', 'chuc_vus.id')
+            ->select('admin_animes.*', 'chuc_vus.ten_chuc_vu')
+            ->where('admin_animes.id', $request->id_admin)
+            ->first();
         return response()->json([
             'obj_admin'  => $user,
         ]);
     }
     public function doiPass(Request $request)
     {
-       $check = Auth::guard('admin')->attempt(['email'=>$request->email,'password' =>$request->old_pass, ]);
+        $check = Auth::guard('admin')->attempt(['email' => $request->email, 'password' => $request->old_pass,]);
         if ($check) {
             $user = Auth::guard('admin')->user();
             $user->update([
-                    'email'                 =>$request->email,
-                    'ho_va_ten'             =>$request->ho_va_ten,
-                    'password'              =>bcrypt($request->new_pass),
-                    'hinh_anh'              =>$request->hinh_anh,
+                'email'                 => $request->email,
+                'ho_va_ten'             => $request->ho_va_ten,
+                'password'              => bcrypt($request->new_pass),
+                'hinh_anh'              => $request->hinh_anh,
             ]);
 
             return response()->json([
@@ -43,8 +76,7 @@ class AdminAnimeController extends Controller
                 'status'    => true,
 
             ]);
-        }
-        else {
+        } else {
             return response()->json([
                 'message'   => 'Mật khẩu cũ không hợp lệ!!',
                 'status'    => 'false'
@@ -55,16 +87,17 @@ class AdminAnimeController extends Controller
     {
         try {
             AdminAnime::where('id', $request->id)
-                    ->update([
-                        'email'                 =>$request->email,
-                        'ho_va_ten'             =>$request->ho_va_ten,
-                        'password'              =>($request->password),
-                        'hinh_anh'              =>$request->hinh_anh,
-                    ]);
+                ->update([
+                    'email'                 => $request->email,
+                    'ho_va_ten'             => $request->ho_va_ten,
+                    'password'              => ($request->password),
+                    'hinh_anh'              => $request->hinh_anh,
+                    'id_chuc_vu'            => $request->id_chuc_vu,
+                ]);
 
             return response()->json([
                 'status'     => true,
-                'ho_ten'     => $request ->ho_va_ten,
+                'ho_ten'     => $request->ho_va_ten,
                 'message'    => 'Cập nhật tài khoản thành công!',
             ]);
         } catch (ExceptionEvent $e) {
@@ -78,38 +111,77 @@ class AdminAnimeController extends Controller
     public function taoAdmin(Request $request)
     {
         try {
+            $id_chuc_nang = 1;
+            $user   = Auth::guard('sanctum')->user(); // Chính là người đang login
+            $user_chuc_vu   = $user->id_chuc_vu;    // Giả sử
+            $check  = PhanQuyen::where('id_chuc_vu', $user_chuc_vu)
+                ->where('id_chuc_nang', $id_chuc_nang)
+                ->first();
+            if (!$check) {
+                return response()->json([
+                    'status'  =>  false,
+                    'message' =>  'Bạn không có quyền chức năng này'
+                ]);
+            }
             AdminAnime::create([
-                'email'                 =>$request->email,
-                'ho_va_ten'             =>$request->ho_va_ten,
-                'password'              =>bcrypt($request->password),
-                'hinh_anh'              =>$request->hinh_anh,
-                ]);
-                return response()->json([
-                    'status'   => true ,
-                    'message'  => 'Bạn thêm admin thành công!',
-                ]);
+                'email'                 => $request->email,
+                'ho_va_ten'             => $request->ho_va_ten,
+                'password'              => bcrypt($request->password),
+                'hinh_anh'              => $request->hinh_anh,
+                'id_chuc_vu'            => $request->id_chuc_vu,
+            ]);
+            return response()->json([
+                'status'   => true,
+                'message'  => 'Bạn thêm admin thành công!',
+            ]);
         } catch (ExceptionEvent $e) {
-                return response()->json([
-                    'status'     => false,
-                    'message'    => 'Xoá admin không thành công!!'
-                ]);
+            return response()->json([
+                'status'     => false,
+                'message'    => 'Xoá admin không thành công!!'
+            ]);
         }
-
     }
 
-     public function timAdmin(Request $request)
+    public function timAdmin(Request $request)
     {
-        $key    = '%'. $request->key . '%';
-        $data   = AdminAnime::select('admin_animes.*')
-                    ->where('ho_va_ten', 'like', $key)
-                    ->get(); // get là ra 1 danh sách
+        $key    = '%' . $request->key . '%';
+        $chuc_vu_admin   = ChucVu::select('chuc_vus.*')
+            ->get(); // get là ra 1  sách
+        $dataAdmin   = AdminAnime::join('chuc_vus', 'id_chuc_vu', 'chuc_vus.id')
+            ->select('admin_animes.*', 'chuc_vus.ten_chuc_vu')
+            ->where('ho_va_ten', 'like', $key)
+            ->paginate(6); // get là ra 1  sách
+        $response = [
+            'pagination' => [
+                'total' => $dataAdmin->total(),
+                'per_page' => $dataAdmin->perPage(),
+                'current_page' => $dataAdmin->currentPage(),
+                'last_page' => $dataAdmin->lastPage(),
+                'from' => $dataAdmin->firstItem(),
+                'to' => $dataAdmin->lastItem()
+            ],
+            'dataAdmin' => $dataAdmin
+        ];
         return response()->json([
-        'admin'  =>  $data,
+            'admin'  =>  $response,
+            'chuc_vu_admin'  =>  $chuc_vu_admin,
         ]);
     }
     public function xoaAdmin($id)
     {
         try {
+            $id_chuc_nang = 1;
+            $user   = Auth::guard('sanctum')->user(); // Chính là người đang login
+            $user_chuc_vu   = $user->id_chuc_vu;    // Giả sử
+            $check  = PhanQuyen::where('id_chuc_vu', $user_chuc_vu)
+                ->where('id_chuc_nang', $id_chuc_nang)
+                ->first();
+            if (!$check) {
+                return response()->json([
+                    'status'  =>  false,
+                    'message' =>  'Bạn không có quyền chức năng này'
+                ]);
+            }
             AdminAnime::where('id', $id)->delete();
 
             return response()->json([
@@ -122,21 +194,32 @@ class AdminAnimeController extends Controller
                 'status'     => false,
                 'message'    => 'Xoá Admin không thành công!!'
             ]);
-
         }
-
     }
 
     public function capnhatAdmin(Request $request)
     {
         try {
+            $id_chuc_nang = 1;
+            $user   = Auth::guard('sanctum')->user(); // Chính là người đang login
+            $user_chuc_vu   = $user->id_chuc_vu;    // Giả sử
+            $check  = PhanQuyen::where('id_chuc_vu', $user_chuc_vu)
+                ->where('id_chuc_nang', $id_chuc_nang)
+                ->first();
+            if (!$check) {
+                return response()->json([
+                    'status'  =>  false,
+                    'message' =>  'Bạn không có quyền chức năng này'
+                ]);
+            }
             AdminAnime::where('id', $request->id)
-                    ->update([
-                        'email'                 =>$request->email,
-                        'ho_va_ten'             =>$request->ho_va_ten,
-                        'password'              =>($request->password),
-                        'hinh_anh'              =>$request->hinh_anh,
-                    ]);
+                ->update([
+                    'email'                 => $request->email,
+                    'ho_va_ten'             => $request->ho_va_ten,
+                    'password'              => ($request->password),
+                    'hinh_anh'              => $request->hinh_anh,
+                    'id_chuc_vu'            => $request->id_chuc_vu,
+                ]);
 
             return response()->json([
                 'status'     => true,
@@ -158,9 +241,9 @@ class AdminAnimeController extends Controller
             $tinh_trang_moi = !$request->tinh_trang;
             //   $tinh_trang_moi là trái ngược của $request->tinh_trangs
             AdminAnime::where('id', $request->id)
-                    ->update([
-                        'tinh_trang'    =>$tinh_trang_moi
-                    ]);
+                ->update([
+                    'tinh_trang'    => $tinh_trang_moi
+                ]);
 
             return response()->json([
                 'status'     => true,
@@ -176,7 +259,7 @@ class AdminAnimeController extends Controller
     }
     public function login(Request $request)
     {
-        $check = Auth::guard('admin')->attempt(['email'=>$request->email,'password' =>$request->password, ]);
+        $check = Auth::guard('admin')->attempt(['email' => $request->email, 'password' => $request->password,]);
         if ($check) {
             $user = Auth::guard('admin')->user();
             return response()->json([
@@ -185,8 +268,7 @@ class AdminAnimeController extends Controller
                 'token'     => $user->createToken('api-token-admin')->plainTextToken,
 
             ]);
-        }
-        else {
+        } else {
             return response()->json([
                 'message'   => 'Đăng Nhập không  thành công!!',
                 'status'    => 'false'
@@ -201,6 +283,7 @@ class AdminAnimeController extends Controller
             'ho_va_ten'     => $request->ho_va_ten,
             'password'      => bcrypt($request->password),
             'hinh_anh'      => $request->hinh_anh,
+            'ngay_sinh'      => $request->ngay_sinh,
         ]);
         return response()->json([
             'message'   => 'Tạo tài khoản thành công!!',
@@ -213,42 +296,39 @@ class AdminAnimeController extends Controller
 
         $user = Auth::guard('sanctum')->user();
 
-        if($user)
-        {
+        if ($user) {
             $agent   = new Agent();
             $device  = $agent->device();
             $os      = $agent->platform();
             $browser = $agent->browser();
             DB::table('personal_access_tokens')
-            ->where('id',$user->currentAccessToken()->id)
-            ->update([
-                'ip'            =>  request()->ip(),
-                'device'        =>  $device,
-                'os'            =>  $os,
-                'trinh_duyet'   =>  $browser
-            ]);
+                ->where('id', $user->currentAccessToken()->id)
+                ->update([
+                    'ip'            =>  request()->ip(),
+                    'device'        =>  $device,
+                    'os'            =>  $os,
+                    'trinh_duyet'   =>  $browser
+                ]);
             return response()->json([
-                'email'      => $user ->email,
-                'ho_ten'     => $user ->ho_va_ten,
-                'hinh_anh'   => $user ->hinh_anh,
-                'id_admin'   => $user ->id,
-                'list'       => $user ->tokens,
+                'email'      => $user->email,
+                'ho_ten'     => $user->ho_va_ten,
+                'hinh_anh'   => $user->hinh_anh,
+                'id_admin'   => $user->id,
+                'list'       => $user->tokens,
 
-            ],200);
-        }
-        else
-        {
+            ], 200);
+        } else {
             return response()->json([
                 'message'   => 'Bạn cần đăng nhập hệ thống !!',
                 'status'    => false
-            ],401);
+            ], 401);
         }
     }
     public function xoatoken($id)
     {
         try {
             DB::table('personal_access_tokens')
-            ->where('id', $id)->delete();
+                ->where('id', $id)->delete();
 
             return response()->json([
                 'status'     => true,
@@ -260,8 +340,6 @@ class AdminAnimeController extends Controller
                 'status'     => false,
                 'message'    => 'Xoá token không thành công!!'
             ]);
-
         }
-
     }
 }
