+++ app/Http/Controllers/KhachHangController.php
@@ -2,9 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\DangKyRequest;
 use App\Http\Requests\DangNhapRequest;
+use App\Jobs\MailQuenMatKhau as JobsMailQuenMatKhau;
+use App\Jobs\MailQueue;
+use App\Mail\KichHoatTaiKhoan;
 use App\Mail\mailQuenMatKhau;
 use App\Models\KhachHang;
+use App\Models\PhanQuyen;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -17,14 +22,39 @@ class KhachHangController extends Controller
 {
     public function getData()
     {
-        $data   = KhachHang::select('khach_hangs.*')
-                         ->get(); // get là ra 1 danh sách
+        $id_chuc_nang = 2;
+        $user   = Auth::guard('sanctum')->user(); // Chính là người đang login
+        $user_chuc_vu   = $user->id_chuc_vu;    // Giả sử
+        $check  = PhanQuyen::where('id_chuc_vu', $user_chuc_vu)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if (!$check) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
+        $dataAdmin   = KhachHang::select('khach_hangs.*')
+            ->paginate(6); // get là ra 1  sách
+
+        $response = [
+            'pagination' => [
+                'total' => $dataAdmin->total(),
+                'per_page' => $dataAdmin->perPage(),
+                'current_page' => $dataAdmin->currentPage(),
+                'last_page' => $dataAdmin->lastPage(),
+                'from' => $dataAdmin->firstItem(),
+                'to' => $dataAdmin->lastItem()
+            ],
+            'dataAdmin' => $dataAdmin
+        ];
         return response()->json([
-            'khach_hang'  =>  $data,
+            'khach_hang'  =>  $response,
         ]);
     }
-    public function getDataProfile(Request $request){
-        $user = KhachHang::where('id',$request->id_khach_hang)->first();
+    public function getDataProfile(Request $request)
+    {
+        $user = KhachHang::where('id', $request->id_khach_hang)->first();
         return response()->json([
             'obj_user'  => $user,
         ]);
@@ -32,38 +62,53 @@ class KhachHangController extends Controller
     public function taoKhachHang(Request $request)
     {
         try {
+            $id_chuc_nang = 2;
+            $user   = Auth::guard('sanctum')->user(); // Chính là người đang login
+            $user_chuc_vu   = $user->id_chuc_vu;    // Giả sử
+            $check  = PhanQuyen::where('id_chuc_vu', $user_chuc_vu)
+                ->where('id_chuc_nang', $id_chuc_nang)
+                ->first();
+            if (!$check) {
+                return response()->json([
+                    'status'  =>  false,
+                    'message' =>  'Bạn không có quyền chức năng này'
+                ]);
+            }
             KhachHang::create([
-                'email'                 =>$request->email,
-                'ho_va_ten'             =>$request->ho_va_ten,
-                'password'              =>bcrypt($request->password),
-                'hinh_anh'              =>$request->hinh_anh,
-                ]);
-                return response()->json([
-                    'status'   => true ,
-                    'message'  => 'Bạn thêm khách hàng thành công!',
-                ]);
+                'email'         => $request->email,
+                'ho_va_ten'     => $request->ho_va_ten,
+                'password'      => bcrypt($request->password),
+                'hinh_anh'      => $request->hinh_anh,
+                'ngay_sinh'     => $request->ngay_sinh,
+                'is_done'       => $request->is_done,
+            ]);
+            return response()->json([
+                'status'   => true,
+                'message'  => 'Bạn thêm khách hàng thành công!',
+            ]);
         } catch (ExceptionEvent $e) {
-                return response()->json([
-                    'status'     => false,
-                    'message'    => 'Xoá khách hàng không thành công!!'
-                ]);
+            return response()->json([
+                'status'     => false,
+                'message'    => 'Xoá khách hàng không thành công!!'
+            ]);
         }
-
     }
     public function doiThongTin(Request $request)
     {
         try {
             KhachHang::where('id', $request->id)
-                    ->update([
-                        'email'                 =>$request->email,
-                        'ho_va_ten'             =>$request->ho_va_ten,
-                        'password'              =>($request->password),
-                        'hinh_anh'              =>$request->hinh_anh,
-                    ]);
+                ->update([
+                    'email'         => $request->email,
+                    'ho_va_ten'     => $request->ho_va_ten,
+                    'password'      => bcrypt($request->password),
+                    'hinh_anh'      => $request->hinh_anh,
+                    'ngay_sinh'     => $request->ngay_sinh,
+                    'is_done'       => $request->is_done,
+                ]);
 
             return response()->json([
                 'status'     => true,
-                'ho_ten_user'=> $request ->ho_va_ten,
+                'ho_ten_user' => $request->ho_va_ten,
                 'message'    => 'Cập nhật tài khoản thành công!',
             ]);
         } catch (ExceptionEvent $e) {
@@ -76,14 +121,14 @@ class KhachHangController extends Controller
     }
     public function doiPass(Request $request)
     {
-       $check = Auth::guard('khach_hang')->attempt(['email'=>$request->email,'password' =>$request->old_pass, ]);
+        $check = Auth::guard('khach_hang')->attempt(['email' => $request->email, 'password' => $request->old_pass,]);
         if ($check) {
             $user = Auth::guard('khach_hang')->user();
             $user->update([
-                    'email'                 =>$request->email,
-                    'ho_va_ten'             =>$request->ho_va_ten,
-                    'password'              =>bcrypt($request->new_pass),
-                    'hinh_anh'              =>$request->hinh_anh,
+                'email'                 => $request->email,
+                'ho_va_ten'             => $request->ho_va_ten,
+                'password'              => bcrypt($request->new_pass),
+                'hinh_anh'              => $request->hinh_anh,
             ]);
 
             return response()->json([
@@ -91,8 +136,7 @@ class KhachHangController extends Controller
                 'status'    => true,
 
             ]);
-        }
-        else {
+        } else {
             return response()->json([
                 'message'   => 'Mật khẩu cũ không hợp lệ!!',
                 'status'    => 'false'
@@ -100,19 +144,43 @@ class KhachHangController extends Controller
         }
     }
 
-     public function timKhachHang(Request $request)
+    public function timKhachHang(Request $request)
     {
-        $key    = '%'. $request->key . '%';
-        $data   = KhachHang::select('khach_hangs.*')
-                    ->where('ho_va_ten', 'like', $key)
-                    ->get(); // get là ra 1 danh sách
-        return response()->json([
-        'khach_hang'  =>  $data,
-        ]);
+        $key    = '%' . $request->key . '%';
+        $dataAdmin   = KhachHang::select('khach_hangs.*')
+            ->where('ho_va_ten', 'like', $key)
+            ->paginate(6); // get là ra 1  sách
+
+            $response = [
+                'pagination' => [
+                    'total' => $dataAdmin->total(),
+                    'per_page' => $dataAdmin->perPage(),
+                    'current_page' => $dataAdmin->currentPage(),
+                    'last_page' => $dataAdmin->lastPage(),
+                    'from' => $dataAdmin->firstItem(),
+                    'to' => $dataAdmin->lastItem()
+                ],
+                'dataAdmin' => $dataAdmin
+            ];
+            return response()->json([
+                'khach_hang'  =>  $response,
+            ]);
     }
     public function xoaKhachHang($id)
     {
         try {
+            $id_chuc_nang = 2;
+            $user   = Auth::guard('sanctum')->user(); // Chính là người đang login
+            $user_chuc_vu   = $user->id_chuc_vu;    // Giả sử
+            $check  = PhanQuyen::where('id_chuc_vu', $user_chuc_vu)
+                ->where('id_chuc_nang', $id_chuc_nang)
+                ->first();
+            if (!$check) {
+                return response()->json([
+                    'status'  =>  false,
+                    'message' =>  'Bạn không có quyền chức năng này'
+                ]);
+            }
             KhachHang::where('id', $id)->delete();
 
             return response()->json([
@@ -125,21 +193,31 @@ class KhachHangController extends Controller
                 'status'     => false,
                 'message'    => 'Xoá khách hàng không thành công!!'
             ]);
-
         }
-
     }
 
     public function capnhatKhachHang(Request $request)
     {
         try {
+            $id_chuc_nang = 2;
+            $user   = Auth::guard('sanctum')->user(); // Chính là người đang login
+            $user_chuc_vu   = $user->id_chuc_vu;    // Giả sử
+            $check  = PhanQuyen::where('id_chuc_vu', $user_chuc_vu)
+                ->where('id_chuc_nang', $id_chuc_nang)
+                ->first();
+            if (!$check) {
+                return response()->json([
+                    'status'  =>  false,
+                    'message' =>  'Bạn không có quyền chức năng này'
+                ]);
+            }
             KhachHang::where('id', $request->id)
-                    ->update([
-                        'email'                 =>$request->email,
-                        'ho_va_ten'             =>$request->ho_va_ten,
-                        'password'              =>($request->password),
-                        'hinh_anh'              =>$request->hinh_anh,
-                    ]);
+                ->update([
+                    'email'                 => $request->email,
+                    'ho_va_ten'             => $request->ho_va_ten,
+                    'password'              => ($request->password),
+                    'hinh_anh'              => $request->hinh_anh,
+                ]);
 
             return response()->json([
                 'status'     => true,
@@ -157,32 +235,32 @@ class KhachHangController extends Controller
 
     public function login(DangNhapRequest $request)
     {
-        $check = Auth::guard('khach_hang')->attempt(['email'=>$request->email,'password' =>$request->password, 'is_done'    => 1 ]);
+        $check = Auth::guard('khach_hang')->attempt(['email' => $request->email, 'password' => $request->password, 'is_done'    => 1]);
         if ($check) {
             $user = Auth::guard('khach_hang')->user();
             return response()->json([
-                'message'   => 'Tài khoản chưa kích hoạt!!',
+                'message'   => 'Đăng nhập thành công!!',
                 'status'    => true,
                 'token'     => $user->createToken('api-token-khach')->plainTextToken,
 
             ]);
-        }
-        else {
+        } else {
             return response()->json([
-                'message'   => 'Đăng Nhập không  thành công!!',
+                'message'   => 'Tài khoản của bạn chưa kích hoạt!!',
                 'status'    => false
             ]);
         }
     }
 
-    public function register(Request $request)
+    public function register(DangKyRequest $request)
     {
         KhachHang::create([
             'email'         => $request->email,
             'ho_va_ten'     => $request->ho_va_ten,
             'password'      => bcrypt($request->password),
             'hinh_anh'      => $request->hinh_anh,
-            'is_done'      => $request->is_done,
+            'ngay_sinh'     => $request->ngay_sinh,
+            'is_done'       => $request->is_done,
 
         ]);
         return response()->json([
@@ -193,7 +271,7 @@ class KhachHangController extends Controller
     public function kiemTraQuenMK(Request $request)
     {
         $check  = KhachHang::where('hash_quen_mat_khau', $request->hash_quen_mat_khau)->first();
-        if($check) {
+        if ($check) {
             return response()->json([
                 'status'            =>   true,
                 'message'           =>   'Vui lòng đặt lại mật khẩu!',
@@ -205,11 +283,57 @@ class KhachHangController extends Controller
             ]);
         }
     }
+    public function kiemTraHashLogin(Request $request)
+    {
+        $khach_hang  = KhachHang::where('hash_kich_hoat', $request->hash_kich_hoat)
+            ->first();
+        if ($khach_hang) {
+            $khach_hang->is_done   =   1;
+            $khach_hang->hash_kich_hoat   =   null;
+            $khach_hang->save();
+            return response()->json([
+                'status'            =>   true,
+                'message'           =>   'Kích hoạt email thành công!',
+                'email'             =>   $khach_hang->email,
+            ]);
+        } else {
+            return response()->json([
+                'status'            =>   false,
+                'message'           =>   'Bạn không được ở đây!',
+            ]);
+        }
+    }
+    public function kichHoatTK(Request $request)
+    {
+        // Gửi lên 1 thằng duy nhất $request->email
+        $khach_hang   = KhachHang::where('email', $request->email)->first();
+        if ($khach_hang) {
+            // Tạo 1 mã hash_kich_hoat
+            $hash_kich_hoat                      =   Str::uuid();
+            $khach_hang->hash_kich_hoat   =   $hash_kich_hoat;
+            $khach_hang->save();
+            // Gửi Email
+            $data['email']  =    $request->email;
+            $data['name']  =    $khach_hang->ho_va_ten;
+            $data['link']  =    'http://localhost:5173/home/kich-hoat-email/' . $hash_kich_hoat;
+            MailQueue::dispatch($data);
+
+            return response()->json([
+                'status'            =>   true,
+                'message'           =>   'Vui lòng kiểm tra email của bạn để kích hoạt!',
+            ]);
+        } else {
+            return response()->json([
+                'status'            =>   false,
+                'message'           =>   'Tài khoản của bạn không tồn tại!',
+            ]);
+        }
+    }
     public function datLaiMK(Request $request)
     {
 
         $khach_hang  = KhachHang::where('hash_quen_mat_khau', $request->hash_quen_mat_khau)->first();
-        if($khach_hang) {
+        if ($khach_hang) {
             $khach_hang->password             =   bcrypt($request->password);
             $khach_hang->hash_quen_mat_khau   =   null;
             $khach_hang->save();
@@ -230,15 +354,16 @@ class KhachHangController extends Controller
     {
         // Gửi lên 1 thằng duy nhất $request->email
         $khach_hang   = KhachHang::where('email', $request->email)->first();
-        if($khach_hang) {
+        if ($khach_hang) {
             // Tạo 1 mã hash_quen_mat_khau
             $hash_pass                      =   Str::uuid();
             $khach_hang->hash_quen_mat_khau   =   $hash_pass;
             $khach_hang->save();
             // Gửi Email
-            $info['name']  =    $khach_hang->ho_va_ten;
-            $info['link']  =    'http://localhost:5173/reset-password/' . $hash_pass;
-            Mail::to($request->email)->send(new mailQuenMatKhau('mail.quen_mat_khau', 'Khôi Phục Tài Khoản Của Bạn', $info));
+            $data['email']  =    $request->email;
+            $data['name']  =    $khach_hang->ho_va_ten;
+            $data['link']  =    'http://localhost:5173/home/reset-password/' . $hash_pass;
+            JobsMailQuenMatKhau::dispatch($data);
             return response()->json([
                 'status'            =>   true,
                 'message'           =>   'Vui lòng kiểm tra email của bạn để đổi lại mật khẩu!',
@@ -256,42 +381,39 @@ class KhachHangController extends Controller
 
         $user = Auth::guard('sanctum')->user();
 
-        if($user)
-        {
+        if ($user) {
             $agent   = new Agent();
             $device  = $agent->device();
             $os      = $agent->platform();
             $browser = $agent->browser();
             DB::table('personal_access_tokens')
-            ->where('id',$user->currentAccessToken()->id)
-            ->update([
-                'ip'            =>  request()->ip(),
-                'device'        =>  $device,
-                'os'            =>  $os,
-                'trinh_duyet'   =>  $browser
-            ]);
+                ->where('id', $user->currentAccessToken()->id)
+                ->update([
+                    'ip'            =>  request()->ip(),
+                    'device'        =>  $device,
+                    'os'            =>  $os,
+                    'trinh_duyet'   =>  $browser
+                ]);
             return response()->json([
-                'email'                => $user ->email,
-                'id_user'              => $user ->id,
-                'ho_ten_user'          => $user ->ho_va_ten,
-                'hinh_anh_user'        => $user ->hinh_anh,
-                'list'                 => $user ->tokens,
+                'email'                => $user->email,
+                'id_user'              => $user->id,
+                'ho_ten_user'          => $user->ho_va_ten,
+                'hinh_anh_user'        => $user->hinh_anh,
+                'list'                 => $user->tokens,
 
-            ],200);
-        }
-        else
-        {
+            ], 200);
+        } else {
             return response()->json([
                 'message'   => 'Bạn cần đăng nhập hệ thống !!',
                 'status'    => false
-            ],401);
+            ], 401);
         }
     }
     public function xoatoken($id)
     {
         try {
             DB::table('personal_access_tokens')
-            ->where('id', $id)->delete();
+                ->where('id', $id)->delete();
 
             return response()->json([
                 'status'     => true,
@@ -303,8 +425,6 @@ class KhachHangController extends Controller
                 'status'     => false,
                 'message'    => 'Xoá Nha token không thành công!!'
             ]);
-
         }
-
     }
 }
